requirements.txt
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-telegram-bot-calendar==1.0.5
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
yarl==1.7.2
